created API for getOutletsByMerchantID and fetching that whether outlet is approved or not through one column is approved through this API, we are Fetching Outlet basic details and adresses like state city and area that are mapped to the address table.

+++ foodandmart/src/main/java/com/jippy/foodandmart/entity/FmMerchantBankDetails.java
@@ -21,7 +21,7 @@ public class FmMerchantBankDetails {
 
     /** Renamed from merchant_id; holds merchantId, driverId, etc. */
     @Column(name = "recipient_id", nullable = false)
-    private Long recipientId;
+    private Integer recipientId;
 
     @Column(name = "account_number", length = 30)
     private String accountNumber;
